foi arrumado o faturamento e os arquivos
