+++ package.json
@@ -19,6 +19,7 @@
     "class-variance-authority": "^0.7.0",
     "clsx": "^2.1.1",
     "dayjs": "^1.11.11",
+    "firebase": "^11.0.1",
     "lucide-react": "^0.390.0",
     "next": "^14.0.5-canary.59",
     "node-fetch": "^3.3.2",
